backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.8.2
